+++ e2e/__tests__/failureDetailsProperty.test.ts
@@ -90,7 +90,7 @@ test('that the failureDetails property is set', () => {
             "actual": "",
             "error": Object {
               "matcherResult": Object {
-                "actual": "Object {
+                "actual": "{
         "p1": "hello",
         "p2": "world",
       }",
@@ -105,7 +105,7 @@ test('that the failureDetails property is set', () => {
       - Snapshot  - 1
       + Received  + 1
 
-        Object {
+        {
           "p1": "hello",
       -   "p2": "sunshine",
       +   "p2": "world",
@@ -123,7 +123,7 @@ test('that the failureDetails property is set', () => {
       - Snapshot  - 1
       + Received  + 1
 
-        Object {
+        {
           "p1": "hello",
       -   "p2": "sunshine",
       +   "p2": "world",
@@ -136,7 +136,7 @@ test('that the failureDetails property is set', () => {
       - Snapshot  - 1
       + Received  + 1
 
-        Object {
+        {
           "p1": "hello",
       -   "p2": "sunshine",
       +   "p2": "world",
@@ -215,7 +215,7 @@ test('that the failureDetails property is set', () => {
         Array [
           Object {
             "matcherResult": Object {
-              "actual": "Object {
+              "actual": "{
         "p1": "hello",
         "p2": "world",
       }",
@@ -230,7 +230,7 @@ test('that the failureDetails property is set', () => {
       - Snapshot  - 1
       + Received  + 1
 
-        Object {
+        {
           "p1": "hello",
       -   "p2": "sunshine",
       +   "p2": "world",